+++ crates/sozo/Cargo.toml
@@ -33,7 +33,6 @@ smol_str.workspace = true
 starknet.workspace = true
 thiserror.workspace = true
 tokio = { version = "1.15.0", features = [ "full" ] }
-toml.workspace = true
 tracing-log = "0.1.3"
 tracing.workspace = true
 url = "2.2.2"
